refactor & feat - knex refactored into abstract class, dbs and controllers implemented
