+++ src/database/poo.sql
@@ -1,4 +1,4 @@
--- Active: 1675168195024@@127.0.0.1@3306
+-- Active: 1675253388925@@127.0.0.1@3306
 CREATE TABLE users (
     id TEXT PRIMARY KEY UNIQUE NOT NULL,
     name TEXT NOT NULL,
